server/requirements.txt
flask==2.3.3
langchain-openai==0.0.2
langchain==0.0.267
langchain-core==0.1.7
python-dotenv==1.0.5
openai>=1.0.0
flask-cors==4.0.0 